# Refactor attribute marshaling to use standard marshalers