… exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/AttributeMarshaler.java
… exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/TraceRequestMarshaler.java